Testing: dumper, migrate data model tests to unit tests
Refactored staticmethod tests into normal tests in order to remove unnecessary pylint comments as well
